Increased test coverage
